requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via realtime
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
black==25.1.0
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
cachetools==5.5.1
    # via streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   streamlit
deprecation==2.1.0
    # via postgrest
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
gotrue==2.11.3
    # via supabase
h11==0.14.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx[http2]==0.28.1
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
narwhals==1.26.0
    # via altair
numpy==2.2.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   black
    #   deprecation
    #   streamlit
pandas==2.2.3
    # via streamlit
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via streamlit
platformdirs==4.3.6
    # via black
postgrest==0.19.3
    # via supabase
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via streamlit
pyarrow==19.0.0
    # via streamlit
pydantic==2.10.6
    # via
    #   gotrue
    #   postgrest
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   realtime
    #   storage3
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.1
    # via pandas
realtime==2.3.0
    # via supabase
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
storage3==0.11.3
    # via supabase
streamlit==1.42.0
    # via -r requirements.in
strenum==0.4.15
    # via supafunc
supabase==2.13.0
    # via -r requirements.in
supafunc==0.9.3
    # via supabase
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   realtime
    #   streamlit
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
websockets==14.2
    # via realtime
yarl==1.18.3
    # via aiohttp
